tun: fix types

+++ src/tun.c
@@ -84,7 +84,7 @@ static const char *debug_payload(uint8_t *buf, size_t buflen)
         return ret;
     }
 
-    int ip_version = (buf[0] >> 4) & 0x0f;
+    uint8_t ip_version = (buf[0] >> 4) & 0x0f;
 
     if (ip_version == 4 && buflen >= 20) {
         // IPv4 packet
@@ -92,11 +92,11 @@ static const char *debug_payload(uint8_t *buf, size_t buflen)
         size_t length = ntohs(*((uint16_t*) &buf[2]));
         struct in_addr *saddr = (struct in_addr *) &buf[12];
         struct in_addr *daddr = (struct in_addr *) &buf[16];
-        int protocol = buf[9];
+        uint8_t protocol = buf[9];
 
         if (protocol == 0x06 || protocol == 0x11) {
-            int sport = ntohs(*((uint16_t*) &buf[20]));
-            int dport = ntohs(*((uint16_t*) &buf[22]));
+            uint16_t sport = ntohs(*((uint16_t*) &buf[20]));
+            uint16_t dport = ntohs(*((uint16_t*) &buf[22]));
 
             snprintf(ret, sizeof(ret), "IPv4/%s, iplen: %zu, %s:%d => %s:%d",
                 protocol_str(protocol), length, str_in4(saddr), sport, str_in4(daddr), dport
@@ -111,11 +111,11 @@ static const char *debug_payload(uint8_t *buf, size_t buflen)
         size_t length = ntohs(*((uint16_t*) &buf[4]));
         struct in6_addr *saddr = (struct in6_addr *) &buf[8];
         struct in6_addr *daddr = (struct in6_addr *) &buf[24];
-        int protocol = buf[6];
+        uint8_t protocol = buf[6];
 
         if (protocol == 0x06 || protocol == 0x11) {
-            int sport = ntohs(*((uint16_t*) &buf[40]));
-            int dport = ntohs(*((uint16_t*) &buf[42]));
+            uint16_t sport = ntohs(*((uint16_t*) &buf[40]));
+            uint16_t dport = ntohs(*((uint16_t*) &buf[42]));
             snprintf(ret, sizeof(ret), "IPv6/%s, iplen: %zu, [%s]:%d => [%s]:%d",
                 protocol_str(protocol), length, str_in6(saddr), sport, str_in6(daddr), dport
             );
